scolisense/scolisense/Python/requirements.txt
pandas>=1.3.0
scikit-learn>=0.24.0
joblib>=1.0.0
pyserial>=3.5
numpy>=1.21.0